requirements.txt
numpy
scipy
scikit-learn
scikit-image
opencv-python
pandas
matplotlib
seaborn
tensorflow-gpu
tensorflow-datasets
imageio
embed
tqdm